Move implementation of e2e_pytest decorator to a separate file
