+++ tests/e2e_test_system.py
@@ -1,0 +1,45 @@
+# Copyright (C) 2021 Intel Corporation
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions
+# and limitations under the License.
+
+import functools
+import pytest
+
+def _generate_e2e_pytest_decorator():
+    try:
+        from e2e.markers.mark_meta import MarkMeta
+    except ImportError:
+        def _e2e_pytest(func):
+            return func
+        return _e2e_pytest
+
+    class Requirements:
+        # Dummy requirement
+        REQ_DUMMY = "Dummy requirement"
+
+    class OTEComponent(MarkMeta):
+        OTE = "ote"
+
+    def _e2e_pytest(func):
+        @pytest.mark.components(OTEComponent.OTE)
+        @pytest.mark.priority_medium
+        @pytest.mark.reqids(Requirements.REQ_DUMMY)
+        @pytest.mark.api_other
+        @functools.wraps(func)
+        def wrapper(*args, **kwargs):
+            return func(*args, **kwargs)
+        return wrapper
+
+    return _e2e_pytest
+
+e2e_pytest = _generate_e2e_pytest_decorator()
